saas_external_domain_resolver/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
psycopg2-binary>=2.9.0